backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
djangorestframework-simplejwt==4.7.2
pillow==10.0.1

